Add new glibc-langpack-en module (for encoding module)
On EL 8 glibc-langpack-en package needs to be installed in
order for the encoding module to work. This language pack
provides the en_US language pack used in the image via the
encoding module. On EL 7 this package doesn't exist and is
not available, thus, don't use that module in the tooling
image.